+++ src/errorHandler.ts
@@ -50,6 +50,13 @@ import {
   AccessTokenExpiredError,
   AccessTokenUnknownError,
   DuplicateRequestError,
+  InvalidCurrencyForProductPrice,
+  InvalidReceiptDataError,
+  UnknownReceiptTransactionError,
+  AppStoreTransactionAlreadyLinked,
+  NoConfiguredAppStoreProduct,
+  StripePaymentMethodError,
+  StripeRequestError,
   IDFormatError,
 } from './errors';
 
@@ -100,6 +107,13 @@ export const ErrorClassDefinitionsMap = {
   2606: UnauthorizedTokenError,
   2607: TokenNotDeleteableError,
   2610: FileTooLargeError,
+  2707: InvalidReceiptDataError,
+  2708: UnknownReceiptTransactionError,
+  2712: StripeRequestError,
+  2713: StripePaymentMethodError,
+  2714: InvalidCurrencyForProductPrice,
+  2716: NoConfiguredAppStoreProduct,
+  2717: AppStoreTransactionAlreadyLinked,
 };
 
 const ErrorClassDefinitionsByErrorMap = {